cleanup, bug fixes, and ui updates
