+++ src/components/style.css
@@ -128,6 +128,28 @@ h1.title {
     width: 145px;
 }
 
+.card-back {
+    background-color: rgba(0, 0, 0, .9) !important;
+    text-align: left !important;
+}
+
+.card-back .hive {
+    opacity: .1 !important;
+}
+
+.energySourcesList {
+    position: absolute !important;
+    left: 0;
+    top: 30px;
+    width: 100%;
+    height: 150px !important;
+    overflow-y: scroll;
+}
+
+.energySourcesList .MuiSvgIcon-root {
+    color: #ffffff !important;
+}
+
 .fab-reaction {
     opacity: 0;
     position: absolute !important;
@@ -143,7 +165,7 @@ h1.title {
 .charged .fab-reaction {
     opacity: 1;
     position: absolute;
-    right: -15px;
+    right: -30px;
     top: 70px;
 }
